Ui updates and bug fixes.
