--- app/src/main/java/com/speakeasy/watsonbarassistant/IngredientDelegate.kt
@@ -1,5 +1,0 @@
-package com.speakeasy.watsonbarassistant
-
-interface  IngredientDelegate {
-    fun addIngredient(name: String)
-}
