+++ app/src/main/java/com/speakeasy/watsonbarassistant/adapter/IngredientsAdapter.kt
@@ -1,9 +1,11 @@
-package com.speakeasy.watsonbarassistant
+package com.speakeasy.watsonbarassistant.adapter
 
 import android.support.v7.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import android.widget.TextView
+import com.speakeasy.watsonbarassistant.Ingredient
+import com.speakeasy.watsonbarassistant.R
 import java.util.*
 
 class IngredientsAdapter(private val ingredientsSet: TreeSet<Ingredient>):
@@ -16,7 +18,7 @@ class IngredientsAdapter(private val ingredientsSet: TreeSet<Ingredient>):
 
     class ViewHolder(val textView: TextView) : RecyclerView.ViewHolder(textView)
 
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): IngredientsAdapter.ViewHolder {
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
         val textView = LayoutInflater.from(parent.context).inflate(R.layout.ingredient_view,
                 parent, false) as TextView
         return ViewHolder(textView)
