+++ app/src/main/java/com/speakeasy/watsonbarassistant/discovery/SearchDiscovery.kt
@@ -6,7 +6,6 @@ import com.ibm.watson.developer_cloud.discovery.v1.Discovery
 import com.ibm.watson.developer_cloud.discovery.v1.model.QueryOptions
 import com.speakeasy.watsonbarassistant.*
 import kotlinx.serialization.json.JSON
-import java.net.URLEncoder
 
 class SearchDiscovery(private val inputListener: CompletedDiscovery):
         AsyncTask<Array<Ingredient>, Void, MutableList<DiscoveryRecipe>>() {
@@ -72,7 +71,7 @@ class SearchDiscovery(private val inputListener: CompletedDiscovery):
             ans += "("
 
             for (element in set){
-                ans += element + ","
+                ans += "$element,"
             }
 
             if (set.count() > 0){
