+++ app/src/main/java/com/speakeasy/watsonbarassistant/extensions/ContextExtensions.kt
@@ -1,0 +1,8 @@
+package com.speakeasy.watsonbarassistant.extensions
+
+import android.content.Context
+import android.widget.Toast
+
+fun Context.toast(message: String, length: Int = Toast.LENGTH_SHORT) {
+    Toast.makeText(this, message, length).show()
+}
