+++ app/src/main/java/com/speakeasy/watsonbarassistant/speech/HandleTtS.kt
@@ -8,7 +8,6 @@ import java.io.IOException
 class HandleTtS: CompletedTtS {
 
     override fun onTaskCompleted(outFile:  String, mediaPlayer: MediaPlayer) {
-
         if (outFile == ERROR_FILE_NAME){
             Log.i("HandlerTtS", "Invalid file returned.")
         } else {
