--- app/src/main/res/drawable/button_bg_round.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<selector xmlns:android="http://schemas.android.com/apk/res/android">
-    <item>
-        <shape android:shape="oval">
-            <solid android:color="@color/colorLightPrimary"/>
-            <size android:width="48dp" android:height="48dp"/>
-        </shape>
-    </item>
-</selector>
