+++ app/src/main/res/layout/activity_main_menu.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainMenu">
+    tools:context=".activity.MainMenu">
 
     <include
         android:id="@+id/toolbar"
