+++ app/src/main/res/layout/activity_recipe_detail.xml
@@ -8,7 +8,7 @@
     <ScrollView
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".RecipeDetail">
+        tools:context=".activity.RecipeDetail">
 
         <LinearLayout
             android:layout_marginBottom="20dp"
@@ -65,6 +65,14 @@
                 android:layout_marginEnd="20dp"
                 android:textColor="@color/primaryText"
                 android:textSize="16sp" />
+
+            <!--Button
+                android:id="@+id/readDescriptionButton"
+                android:layout_marginStart="15dp"
+                android:layout_marginEnd="15dp"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+            android:text="@string/speech_button" /-->
 
             <TextView
                 android:layout_width="wrap_content"
