+++ app/src/main/res/layout/activity_shopping_cart.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".ShoppingCart">
+    tools:context=".activity.ShoppingCart">
 
     <TextView
         android:layout_width="wrap_content"
