+++ app/src/main/res/layout/fragment_home_tab.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".HomeTab"
+    tools:context=".fragment.HomeTab"
     android:orientation="vertical"
     android:layout_marginTop="20dp">
