+++ app/src/main/res/layout/fragment_ingredient_tab.xml
@@ -9,7 +9,7 @@
     android:focusable="true"
     android:focusableInTouchMode="true"
     android:fitsSystemWindows="false"
-    tools:context=".IngredientsTab">
+    tools:context=".fragment.IngredientsTab">
 
     <android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
         xmlns:app="http://schemas.android.com/apk/res-auto"
@@ -17,7 +17,7 @@
         android:id="@+id/relativeLayout"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".IngredientsTab">
+        tools:context=".fragment.IngredientsTab">
 
         <EditText
             android:id="@+id/ingredientInputView"
