+++ app/src/main/res/layout/fragment_my_favorites_tab.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MyFavoritesTab"
+    tools:context=".fragment.FavoritesTab"
     android:orientation="vertical">
 
     <android.support.v7.widget.RecyclerView
