--- app/src/main/res/layout/home_buttons_view.xml
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:orientation="horizontal"
-    android:gravity="center_horizontal"
-    android:layout_marginBottom="20dp"
-    android:layout_gravity="center_horizontal"
-    xmlns:android="http://schemas.android.com/apk/res/android">
-
-        <ImageButton
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:contentDescription="@string/shopping_cart"
-            android:background="@drawable/button_bg_round"
-            android:src="@drawable/ic_shopping_cart_black_24dp"
-            android:layout_marginEnd="100dp"/>
-
-        <ImageButton
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:contentDescription="@string/ingredient_page_description"
-            android:background="@drawable/button_bg_round"
-            android:src="@drawable/ic_search_black_24dp" />
-
-    </LinearLayout>
