+++ app/src/main/res/menu/main_toolbar_menu.xml
@@ -3,7 +3,7 @@
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".MainMenu"
+    tools:context=".activity.MainMenu"
     android:layout_height="wrap_content"
     android:layout_width="match_parent">
