+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,8 +1,4 @@
-<<<<<<< Updated upstream
-#Thu Sep 27 00:09:13 EDT 2018
-=======
 #Sun Oct 21 01:18:31 EDT 2018
->>>>>>> Stashed changes
 distributionBase=GRADLE_USER_HOME
 distributionPath=wrapper/dists
 zipStoreBase=GRADLE_USER_HOME
